Fix 1 5-flip_bits.c

+++ 0x14-bit_manipulation/5-flip_bits.c
@@ -10,7 +10,7 @@
 unsigned int flip_bits(unsigned long int n, unsigned long int m)
 {
 	unsigned int count = 0;
-	unsigned int rem = n ^ m;
+	unsigned long int rem = n ^ m;
 
 	while (rem != 0)
 	{
